requirements.txt
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
iniconfig==2.3.0
lxml==6.0.2
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
pypandoc==1.16.2
pytest==9.0.2
PyYAML==6.0.3
requests==2.32.5
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
